// lib/projects/types.ts
export type Project = {
  id: string; title: string; slug: string; shortDescription: string; description: string;
  location: string; temperature: string; capacity: string; category: string; images: string[];
  coverImage: string; isVisible: boolean; isFeatured: boolean; order: number; createdAt: string; updatedAt: string;
};